requirements.txt
boto3==1.28.78
botocore==1.31.78
python-decouple==3.8